some small refactoring
