+++ modules/back-end/src/Application/FeatureFlags/PatchFeatureFlag.cs
@@ -1,6 +1,7 @@
-using Application.Users;
-using Microsoft.AspNetCore.JsonPatch;
 using Domain.AuditLogs;
+using Application.Users;
+using Application.Bases.Models;
+using Microsoft.AspNetCore.JsonPatch;
 
 namespace Application.FeatureFlags;
